+++ locust/env.py
@@ -170,6 +170,7 @@ class Environment:
         self,
         host="",
         port=8089,
+        web_base_path: str | None = None,
         web_login: bool = False,
         tls_cert: str | None = None,
         tls_key: str | None = None,
@@ -204,6 +205,7 @@ class Environment:
             delayed_start=delayed_start,
             userclass_picker_is_active=userclass_picker_is_active,
             build_path=build_path,
+            web_base_path=web_base_path,
         )
         return self.web_ui
 
@@ -298,8 +300,7 @@ class Environment:
     def _validate_shape_class_instance(self):
         if self.shape_class is not None and not isinstance(self.shape_class, LoadTestShape):
             raise ValueError(
-                "shape_class should be instance of LoadTestShape or subclass LoadTestShape, but got: %s"
-                % self.shape_class
+                f"shape_class should be instance of LoadTestShape or subclass LoadTestShape, but got: {self.shape_class}"
             )
 
     @property