+++ locust/web.py
@@ -11,10 +11,11 @@ from io import StringIO
 from itertools import chain
 from json import dumps
 from time import time
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, TypedDict
 
 import gevent
 from flask import (
+    Blueprint,
     Flask,
     Response,
     jsonify,
@@ -40,7 +41,7 @@ from .runners import STATE_MISSING, STATE_RUNNING, MasterRunner
 from .stats import StatsCSV, StatsCSVFileWriter, StatsErrorDict, sort_stats
 from .user.inspectuser import get_ratio
 from .util.cache import memoize
-from .util.date import format_utc_timestamp
+from .util.date import format_safe_timestamp
 from .util.timespan import parse_timespan
 
 if TYPE_CHECKING:
@@ -51,6 +52,36 @@ logger = logging.getLogger(__name__)
 greenlet_exception_handler = greenlet_exception_logger(logger)
 
 DEFAULT_CACHE_TIME = 2.0
+
+
+class InputField(TypedDict, total=False):
+    label: str
+    name: str
+    type: str | None
+    default_value: bool | None
+    choices: list[str] | None
+    is_secret: bool | None
+    is_required: bool | None
+
+
+class CustomForm(TypedDict, total=False):
+    inputs: list[InputField] | None
+    callback_url: str
+    submit_button_text: str | None
+
+
+class AuthProvider(TypedDict, total=False):
+    label: str | None
+    callback_url: str
+    icon_url: str | None
+
+
+class AuthArgs(TypedDict, total=False):
+    custom_form: CustomForm
+    auth_providers: list[AuthProvider]
+    username_password_callback: str
+    error: str
+    info: str
 
 
 class WebUI:
@@ -84,7 +115,7 @@ class WebUI:
     """Arguments used to render index.html for the web UI. Must be used with custom templates
     extending index.html."""
 
-    auth_args: dict[str, Any]
+    auth_args: AuthArgs
     """Arguments used to render auth.html for the web UI auth page. Must be used when configuring auth"""
 
     def __init__(
@@ -92,6 +123,7 @@ class WebUI:
         environment: Environment,
         host: str,
         port: int,
+        web_base_path: str | None = None,
         web_login: bool = False,
         tls_cert: str | None = None,
         tls_key: str | None = None,
@@ -133,20 +165,21 @@ class WebUI:
         self.auth_args = {}
         self.app.template_folder = build_path or DEFAULT_BUILD_PATH
         self.app.static_url_path = "/assets/"
+
+        app_blueprint = Blueprint("locust", __name__, url_prefix=web_base_path)
         # ensures static js files work on Windows
         mimetypes.add_type("application/javascript", ".js")
-
         if self.web_login:
             self._login_manager = LoginManager()
             self._login_manager.init_app(self.app)
-            self._login_manager.login_view = "login"
+            self._login_manager.login_view = "locust.login"
 
         if environment.runner:
             self.update_template_args()
         if not delayed_start:
             self.start()
 
-        @app.errorhandler(Exception)
+        @app_blueprint.errorhandler(Exception)
         def handle_exception(error):
             error_message = str(error)
             error_code = getattr(error, "code", 500)
@@ -156,7 +189,7 @@ class WebUI:
             )
             return make_response(error_message, error_code)
 
-        @app.route("/assets/<path:path>")
+        @app_blueprint.route("/assets/<path:path>")
         def send_assets(path):
             directory = (
                 os.path.join(self.app.template_folder, "assets")
@@ -166,7 +199,7 @@ class WebUI:
 
             return send_from_directory(directory, path)
 
-        @app.route("/")
+        @app_blueprint.route("/")
         @self.auth_required_if_enabled
         def index() -> str | Response:
             if not environment.runner:
@@ -175,7 +208,7 @@ class WebUI:
 
             return render_template("index.html", template_args=self.template_args)
 
-        @app.route("/swarm", methods=["POST"])
+        @app_blueprint.route("/swarm", methods=["POST"])
         @self.auth_required_if_enabled
         def swarm() -> Response:
             assert request.method == "POST"
@@ -289,7 +322,7 @@ class WebUI:
             else:
                 return jsonify({"success": False, "message": "No runner", "host": environment.host})
 
-        @app.route("/stop")
+        @app_blueprint.route("/stop")
         @self.auth_required_if_enabled
         def stop() -> Response:
             if self._swarm_greenlet is not None:
@@ -299,7 +332,7 @@ class WebUI:
                 environment.runner.stop()
             return jsonify({"success": True, "message": "Test stopped"})
 
-        @app.route("/stats/reset")
+        @app_blueprint.route("/stats/reset")
         @self.auth_required_if_enabled
         def reset_stats() -> str:
             environment.events.reset_stats.fire()
@@ -308,7 +341,7 @@ class WebUI:
                 environment.runner.exceptions = {}
             return "ok"
 
-        @app.route("/stats/report")
+        @app_blueprint.route("/stats/report")
         @self.auth_required_if_enabled
         def stats_report() -> Response:
             theme = request.args.get("theme", "")
@@ -319,17 +352,25 @@ class WebUI:
             )
             if request.args.get("download"):
                 res = app.make_response(res)
-                res.headers["Content-Disposition"] = f"attachment;filename=report_{time()}.html"
+                host = f"_{self.environment.host}" if self.environment.host else ""
+                res.headers["Content-Disposition"] = (
+                    f"attachment;filename=Locust_{format_safe_timestamp(self.environment.stats.start_time)}_"
+                    + f"{self.environment.locustfile}{host}.html"
+                )
             return res
 
         def _download_csv_suggest_file_name(suggest_filename_prefix: str) -> str:
             """Generate csv file download attachment filename suggestion.
 
             Arguments:
-            suggest_filename_prefix: Prefix of the filename to suggest for saving the download. Will be appended with timestamp.
+            suggest_filename_prefix: Prefix of the filename to suggest for saving the download.
+            Will be appended with timestamp.
             """
-
-            return f"{suggest_filename_prefix}_{time()}.csv"
+            host = f"_{self.environment.host}" if self.environment.host else ""
+            return (
+                f"Locust_{format_safe_timestamp(self.environment.stats.start_time)}_"
+                + f"{self.environment.locustfile}{host}_{suggest_filename_prefix}.csv"
+            )
 
         def _download_csv_response(csv_data: str, filename_prefix: str) -> Response:
             """Generate csv file download response with 'csv_data'.
@@ -346,7 +387,7 @@ class WebUI:
             )
             return response
 
-        @app.route("/stats/requests/csv")
+        @app_blueprint.route("/stats/requests/csv")
         @self.auth_required_if_enabled
         def request_stats_csv() -> Response:
             data = StringIO()
@@ -354,7 +395,7 @@ class WebUI:
             self.stats_csv_writer.requests_csv(writer)
             return _download_csv_response(data.getvalue(), "requests")
 
-        @app.route("/stats/requests_full_history/csv")
+        @app_blueprint.route("/stats/requests_full_history/csv")
         @self.auth_required_if_enabled
         def request_stats_full_history_csv() -> Response:
             options = self.environment.parsed_options
@@ -372,7 +413,7 @@ class WebUI:
 
             return make_response("Error: Server was not started with option to generate full history.", 404)
 
-        @app.route("/stats/failures/csv")
+        @app_blueprint.route("/stats/failures/csv")
         @self.auth_required_if_enabled
         def failures_stats_csv() -> Response:
             data = StringIO()
@@ -380,7 +421,7 @@ class WebUI:
             self.stats_csv_writer.failures_csv(writer)
             return _download_csv_response(data.getvalue(), "failures")
 
-        @app.route("/stats/requests")
+        @app_blueprint.route("/stats/requests")
         @self.auth_required_if_enabled
         @memoize(timeout=DEFAULT_CACHE_TIME, dynamic_timeout=True)
         def request_stats() -> Response:
@@ -450,7 +491,7 @@ class WebUI:
 
             return jsonify(report)
 
-        @app.route("/exceptions")
+        @app_blueprint.route("/exceptions")
         @self.auth_required_if_enabled
         def exceptions() -> Response:
             return jsonify(
@@ -467,7 +508,7 @@ class WebUI:
                 }
             )
 
-        @app.route("/exceptions/csv")
+        @app_blueprint.route("/exceptions/csv")
         @self.auth_required_if_enabled
         def exceptions_csv() -> Response:
             data = StringIO()
@@ -475,7 +516,7 @@ class WebUI:
             self.stats_csv_writer.exceptions_csv(writer)
             return _download_csv_response(data.getvalue(), "exceptions")
 
-        @app.route("/tasks")
+        @app_blueprint.route("/tasks")
         @self.auth_required_if_enabled
         def tasks() -> dict[str, dict[str, dict[str, float]]]:
             runner = self.environment.runner
@@ -495,24 +536,25 @@ class WebUI:
             }
             return task_data
 
-        @app.route("/logs")
+        @app_blueprint.route("/logs")
         @self.auth_required_if_enabled
         def logs():
             return jsonify({"master": get_logs(), "workers": self.environment.worker_logs})
 
-        @app.route("/login")
+        @app_blueprint.route("/login")
         def login():
             if not self.web_login:
-                return redirect(url_for("index"))
+                return redirect(url_for("locust.index"))
 
             self.auth_args["error"] = session.get("auth_error", None)
+            self.auth_args["info"] = session.get("auth_info", None)
 
             return render_template_from(
                 "auth.html",
                 auth_args=self.auth_args,
             )
 
-        @app.route("/user", methods=["POST"])
+        @app_blueprint.route("/user", methods=["POST"])
         def update_user():
             assert request.method == "POST"
 
@@ -520,6 +562,8 @@ class WebUI:
             self.environment.update_user_class(user_settings)
 
             return {}, 201
+
+        app.register_blueprint(app_blueprint)
 
     @property
     def login_manager(self):
@@ -577,6 +621,7 @@ class WebUI:
             if self.web_login:
                 try:
                     session["auth_error"] = None
+                    session["auth_info"] = None
                     return login_required(view_func)(*args, **kwargs)
                 except Exception as e:
                     return f"Locust auth exception: {e} See https://docs.locust.io/en/stable/extending-locust.html#adding-authentication-to-the-web-ui for configuring authentication."