+++ locust/webui/src/pages/Auth.tsx
@@ -6,6 +6,7 @@ import Logo from 'assets/Logo';
 import CustomInput from 'components/Form/CustomInput';
 import PasswordField from 'components/Form/PasswordField';
 import DarkLightToggle from 'components/Layout/Navbar/DarkLightToggle';
+import Markdown from 'components/Markdown/Markdown';
 import useCreateTheme from 'hooks/useCreateTheme';
 import { IAuthArgs } from 'types/auth.types';
 
@@ -13,6 +14,7 @@ export default function Auth({
   authProviders,
   customForm,
   error,
+  info,
   usernamePasswordCallback,
 }: IAuthArgs) {
   const theme = useCreateTheme();
@@ -36,6 +38,7 @@ export default function Auth({
           rowGap: 4,
           boxShadow: 24,
           borderRadius: 4,
+          width: 400,
           border: '3px solid black',
           p: 4,
         }}
@@ -48,6 +51,7 @@ export default function Auth({
             <Box sx={{ display: 'flex', flexDirection: 'column', rowGap: 2 }}>
               <TextField label='Username' name='username' />
               <PasswordField />
+              {info && <Alert severity='info'>{<Markdown content={info} />}</Alert>}
               {error && <Alert severity='error'>{error}</Alert>}
               <Button size='large' type='submit' variant='contained'>
                 Login
@@ -61,6 +65,7 @@ export default function Auth({
               {customForm.inputs.map((inputProps, index) => (
                 <CustomInput key={`custom-form-input-${index}`} {...inputProps} />
               ))}
+              {info && <Alert severity='info'>{<Markdown content={info} />}</Alert>}
               {error && <Alert severity='error'>{error}</Alert>}
               <Button size='large' type='submit' variant='contained'>
                 {customForm.submitButtonText || 'Login'}